[refactor] utils pkg
