--- internal/utils/conv.go
@@ -1,159 +1,0 @@
-package utils
-
-import (
-	"strings"
-	"unicode"
-)
-
-// ToUpperCamelCase converts a string into UpperCamelCase format,
-// ensuring the result is safe for use in Go identifiers.
-//
-// Examples:
-//
-//	ToUpperCamelCase("user_id")         → "UserId"
-//	ToUpperCamelCase("user-name")       → "UserName"
-//	ToUpperCamelCase("1type")           → "X1type"
-//	ToUpperCamelCase("full#access")     → "FullAccess"
-//	ToUpperCamelCase("!@#special-case") → "XxxSpecialCase"
-func ToUpperCamelCase(s string) string {
-	res := ToSafeName(toCamelCase(s))
-	return strings.ToUpper(res[:1]) + res[1:]
-}
-
-// ToLowerCamelCase converts a string into lowerCamelCase format,
-// ensuring the result is safe for use in Go identifiers.
-//
-// Examples:
-//
-//	ToLowerCamelCase("user_id")  → "userId"
-//	ToLowerCamelCase("Type")     → "type"
-//	ToLowerCamelCase("1invalid") → "x1invalid"
-func ToLowerCamelCase(s string) string {
-	res := ToSafeName(toCamelCase(s))
-	return strings.ToLower(res[:1]) + res[1:]
-}
-
-// ToLowerInlineCase converts a string to lowercase without underscores.
-//
-// Examples:
-//
-//	ToLowerInlineCase("user_id")          → "userid"
-//	ToLowerInlineCase("snake_case_value") → "snakecasevalue"
-func ToLowerInlineCase(s string) string {
-	res := strings.ReplaceAll(ToSafeName(s), "_", "")
-	return strings.ToLower(res)
-}
-
-// ToUpperinlineCase converts a string to uppercase without underscores.
-//
-// Examples:
-//
-//	ToUpperinlineCase("user_id")   → "USERID"
-//	ToUpperinlineCase("Api_Token") → "APITOKEN"
-func ToUpperinlineCase(s string) string {
-	res := strings.ReplaceAll(ToSafeName(s), "_", "")
-	return strings.ToUpper(res)
-}
-
-// ToSafeName sanitizes any string into a Go-safe identifier:
-// - replaces invalid characters with underscores
-// - prefixes with 'x' if it starts with a number or is a reserved keyword
-//
-// Examples:
-//
-//	ToSafeName("1test")       → "x1test"
-//	ToSafeName("type")        → "xtype"
-//	ToSafeName("hello_world") → "hello_world"
-//	ToSafeName("$$$abc")      → "abc"
-func ToSafeName(s string) string {
-	s = strings.TrimFunc(s, func(r rune) bool {
-		return (r < 'A' || r > 'Z') &&
-			(r < 'a' || r > 'z') &&
-			(r < '0' || r > '9')
-	})
-	var b strings.Builder
-	for _, r := range s {
-		switch {
-		case (r >= 'A' && r <= 'Z') ||
-			(r >= 'a' && r <= 'z') ||
-			(r >= '0' && r <= '9'):
-			b.WriteRune(r)
-		default:
-			b.WriteRune('_')
-		}
-	}
-	switch {
-	case b.Len() == 0:
-		return "xxx"
-	case unicode.IsDigit(rune(b.String()[0])) || reservedWords[strings.ToLower(b.String())]:
-		return "x" + b.String()
-	default:
-		return b.String()
-	}
-}
-
-var reservedWords = map[string]bool{
-	"break":       true,
-	"continue":    true,
-	"return":      true,
-	"fallthrough": true,
-	"goto":        true,
-
-	"if":      true,
-	"else":    true,
-	"for":     true,
-	"range":   true,
-	"switch":  true,
-	"case":    true,
-	"default": true,
-	"select":  true,
-
-	"var":       true,
-	"const":     true,
-	"type":      true,
-	"struct":    true,
-	"interface": true,
-	"map":       true,
-	"chan":      true,
-	"func":      true,
-	"package":   true,
-	"import":    true,
-	"defer":     true,
-
-	"any": true,
-
-	"go": true,
-}
-
-func toCamelCase(s string) string {
-	var (
-		res     strings.Builder
-		capNext = true
-	)
-
-	for _, r := range s {
-		switch {
-		case r == '_' || r == '-' || r == '#':
-			capNext = true
-		case capNext:
-			res.WriteRune(unicode.ToUpper(r))
-			capNext = false
-		default:
-			res.WriteRune(r)
-		}
-	}
-	return res.String()
-}
-
-// IsFloatType проверяет является ли тип float
-func IsFloatType(goType string) bool {
-	return goType == "float32" || goType == "float64"
-}
-
-// Slice ...
-func Slice(s string, start int) string {
-	if start >= len(s) {
-		return ""
-	}
-	return s[start:]
-}
