--- internal/utils/set.go
@@ -1,13 +1,0 @@
-package utils
-
-import "slices"
-
-// AvailableKeys ...
-func AvailableKeys(m map[string]bool) []string {
-	keys := make([]string, 0, len(m))
-	for k := range m {
-		keys = append(keys, k)
-	}
-	slices.Sort(keys)
-	return keys
-}
